refactor

+++ src/web_templates.h
@@ -39,7 +39,7 @@ void web_handle_notFound() {
   web_send_text(404, "text/plain", "Not Found");
 }
 
-void web_send_page(String title, String body, int refresh = 0) {
+void web_send_page(String title, String body, uint16_t refresh = 0) {
   String html = "<html><head><meta charset=\"utf-8\">";
   html += "<meta http-equiv=\"Content-Type\" content=\"text/html; charset=UTF-8\">";
   html += "<meta http-equiv=\"Cache-Control\" content=\"no-cache\">";
